moved code from lib to ind-cca
