+++ libcrux-ml-kem/src/kyber512.rs
@@ -1,4 +1,4 @@
-use super::{constants::*, *};
+use super::{constants::*, ind_cca::*, *};
 
 // Kyber 512 parameters
 const RANK_512: usize = 2;
@@ -40,7 +40,7 @@ pub type MlKem512PublicKey = MlKemPublicKey<CPA_PKE_PUBLIC_KEY_SIZE_512>;
 ///
 /// Returns `Some(public_key)` if valid, and `None` otherwise.
 pub fn validate_public_key(public_key: MlKem512PublicKey) -> Option<MlKem512PublicKey> {
-    if super::validate_public_key::<
+    if super::ind_cca::validate_public_key::<
         RANK_512,
         RANKED_BYTES_PER_RING_ELEMENT_512,
         CPA_PKE_PUBLIC_KEY_SIZE_512,
@@ -75,7 +75,7 @@ pub fn encapsulate(
     MlKemCiphertext<CPA_PKE_CIPHERTEXT_SIZE_512>,
     MlKemSharedSecret,
 ) {
-    super::encapsulate::<
+    super::ind_cca::encapsulate::<
         RANK_512,
         CPA_PKE_CIPHERTEXT_SIZE_512,
         CPA_PKE_PUBLIC_KEY_SIZE_512,
@@ -97,7 +97,7 @@ pub fn decapsulate(
     secret_key: &MlKemPrivateKey<SECRET_KEY_SIZE_512>,
     ciphertext: &MlKemCiphertext<CPA_PKE_CIPHERTEXT_SIZE_512>,
 ) -> [u8; SHARED_SECRET_SIZE] {
-    super::decapsulate::<
+    super::ind_cca::decapsulate::<
         RANK_512,
         SECRET_KEY_SIZE_512,
         CPA_PKE_SECRET_KEY_SIZE_512,
