+++ libcrux-ml-kem/src/kyber768.rs
@@ -1,4 +1,4 @@
-use super::{constants::*, *};
+use super::{constants::*, ind_cca::*, *};
 
 // Kyber 768 parameters
 const RANK_768: usize = 3;
@@ -41,7 +41,7 @@ pub type MlKem768PublicKey = MlKemPublicKey<CPA_PKE_PUBLIC_KEY_SIZE_768>;
 ///
 /// Returns `Some(public_key)` if valid, and `None` otherwise.
 pub fn validate_public_key(public_key: MlKem768PublicKey) -> Option<MlKem768PublicKey> {
-    if super::validate_public_key::<
+    if super::ind_cca::validate_public_key::<
         RANK_768,
         RANKED_BYTES_PER_RING_ELEMENT_768,
         CPA_PKE_PUBLIC_KEY_SIZE_768,
@@ -57,7 +57,7 @@ pub fn validate_public_key(public_key: MlKem768PublicKey) -> Option<MlKem768Publ
 pub fn generate_key_pair(
     randomness: [u8; KEY_GENERATION_SEED_SIZE],
 ) -> MlKemKeyPair<SECRET_KEY_SIZE_768, CPA_PKE_PUBLIC_KEY_SIZE_768> {
-    generate_keypair::<
+    super::ind_cca::generate_keypair::<
         RANK_768,
         CPA_PKE_SECRET_KEY_SIZE_768,
         SECRET_KEY_SIZE_768,
@@ -76,7 +76,7 @@ pub fn encapsulate(
     MlKemCiphertext<CPA_PKE_CIPHERTEXT_SIZE_768>,
     MlKemSharedSecret,
 ) {
-    super::encapsulate::<
+    super::ind_cca::encapsulate::<
         RANK_768,
         CPA_PKE_CIPHERTEXT_SIZE_768,
         CPA_PKE_PUBLIC_KEY_SIZE_768,
@@ -98,7 +98,7 @@ pub fn decapsulate(
     secret_key: &MlKemPrivateKey<SECRET_KEY_SIZE_768>,
     ciphertext: &MlKemCiphertext<CPA_PKE_CIPHERTEXT_SIZE_768>,
 ) -> [u8; SHARED_SECRET_SIZE] {
-    super::decapsulate::<
+    super::ind_cca::decapsulate::<
         RANK_768,
         SECRET_KEY_SIZE_768,
         CPA_PKE_SECRET_KEY_SIZE_768,
